Use manually-created Signon bearer tokens
This is a partial revert of PR #202. Before we can use the bearer
tokens created by the SecretsManager Lambdas, we first need the
Signon PRs #1623 (Bootstrapping task) and #1617 (Bearer token API)
to be merged.

I thought we could get around this by using an image
of signon based on those two branches, but we're seeing rate limits
pulling the image from DockerHub - this is a quicker fix to the default
cluster.

I'll revert this PR once we're ready to use the autogenerated bearer
tokens.