Provide content schemas to apps
This commit shows how we can provide content schemas to apps running in ECS, via the use of Concourse, S3 and a sidecar container. Concourse is used to ensure that the latest content schemas from the GitHub repo are synced to an S3 bucket, which provides an extra layer of redundancy should there be any issues with GitHub (currently call `git clone` to get the schemas, which can fail on the semi-frequent times when GitHub is unavailable). Next, a sidecar container (per service) spins up, pulls the content schemas from S3 and writes them to a transient volume, which is mounted to both the sidecar container and the app container, at the path that the app expects content schemas to reside at. Finally, the app container spins up and can access the content schemas.