Redeploy apps when secrets are rotated
This adds an s3-iam resource that gives access to S3 buckets
in our AWS account (as opposed to the AWS account owned by
the team behind Big Concourse).

The frontend-deploy-event-trigger Concourse job will redeploy
the frontend app whenever a new object is placed in the frontend
bucket in deploy-events-((workspace)).

The intent is for the Secret Rotation Lambda to place a timestamped
file into the bucket whenever a secret an app uses is rotated,
triggering the app to be redeployed and the env var secrets for the
app's tasks to be refreshed from SecretsManager.